Allow subclassed devices to *add* rather than *replace* FuUdevDevice open flags
Old versions of fwupd used to start with OPEN_READ|OPEN_WRITE and then plugins
would set the new flags to be correct. This doesn't work very well when there
are subclassed devices that want to *inherit* flags, rather than having to set
all the same flags as the baseclass.